Next iteration of IndicesService improvements.

+++ core/src/main/java/org/elasticsearch/index/IndexService.java
@@ -410,15 +410,13 @@ public final class IndexService extends AbstractIndexComponent implements IndexC
 
 
     private void onShardClose(ShardLock lock, boolean ownsShard) {
-        if (deleted.get()) { // we remove that shards content if this index has been deleted
+        if (deleted.get() && ownsShard) { // we remove that shards content if this index has been deleted and we own the shard
             try {
-                if (ownsShard) {
-                    try {
-                        eventListener.beforeIndexShardDeleted(lock.getShardId(), indexSettings.getSettings());
-                    } finally {
-                        shardStoreDeleter.deleteShardStore("delete index", lock, indexSettings);
-                        eventListener.afterIndexShardDeleted(lock.getShardId(), indexSettings.getSettings());
-                    }
+                try {
+                    eventListener.beforeIndexShardDeleted(lock.getShardId(), indexSettings.getSettings());
+                } finally {
+                    shardStoreDeleter.deleteShardStore("delete index", lock, indexSettings);
+                    eventListener.afterIndexShardDeleted(lock.getShardId(), indexSettings.getSettings());
                 }
             } catch (IOException e) {
                 shardStoreDeleter.addPendingDelete(lock.getShardId(), indexSettings);
